-доделал sharedpref

+++ app/src/main/java/com/example/a19mart/db/Node.kt
@@ -13,5 +13,5 @@ data class Node(
     val address: String,
     val children: MutableList<Node> = mutableListOf(),
     @ColumnInfo(name = "parent")
-    var parentId: Int? = null
+    var parentId: Long? = null
 )
